+++ src/nl/hannahsten/texifyidea/index/stub/LatexEnvironmentStubElementType.kt
@@ -6,6 +6,7 @@ import nl.hannahsten.texifyidea.index.NewLabelsIndex
 import nl.hannahsten.texifyidea.psi.LatexEnvironment
 import nl.hannahsten.texifyidea.psi.getEnvironmentName
 import nl.hannahsten.texifyidea.psi.getLabelFromOptionalParameter
+import nl.hannahsten.texifyidea.psi.getLabelFromRequiredParameter
 import nl.hannahsten.texifyidea.psi.impl.LatexEnvironmentImpl
 import java.io.IOException
 
@@ -15,7 +16,7 @@ open class LatexEnvironmentStubElementType(debugName: String) : IStubElementType
 
     override fun createStub(psi: LatexEnvironment, parentStub: StubElement<*>): LatexEnvironmentStub {
         val envName = psi.getEnvironmentName()
-        val directLabel = psi.getLabelFromOptionalParameter()
+        val directLabel = psi.getLabelFromOptionalParameter() ?: psi.getLabelFromRequiredParameter()
         // we only record the label if it is specified as an optional parameter
         // If it is contained in some `\label{}` command inside the environment, it will be indexed separately via the command
         // If we record it here, we will have duplicates in the index